requirements.txt
google-cloud-bigquery==3.11.0
google-cloud-documentai==2.16.0
google-cloud-storage==2.9.0